% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_categorical.R
\name{autoinput_categorical}
\alias{autoinput_categorical}
\title{autoinput_categorical}
\usage{
autoinput_categorical(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be factor)}
}
\value{
Select input for categorical variables
}
\description{
Automatically genrates select inputs for all factor columns in a dataframe.
}
\details{
Based on the factor columns on a dataframe, this function generates all the select inputs. 
         The possible values correspond to the levels in each factor column.
}
\examples{
print('anda')
}
