% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_date.R
\name{autoinput_date}
\alias{autoinput_date}
\title{autoinput_date}
\usage{
autoinput_date(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be date)}
}
\value{
dateInput() for all the date variables.
}
\description{
Automatically genrates date inputs for all date variables in a dataframe.
}
\details{
Based on the date columns on a dataframe, this function generates all the date inputs. 
         The min/max value on the dataframe is the min/max value on the date input.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
load('data/customers.rda')
ui <- fluidPage(
  h3('UI (inputs)'),
  autoinput_date(.df = customers),
  h3('Server (output)'),
  verbatimTextOutput(outputId = 'values')
)
server <- function(input, output) {
  output$values <- reactive({
    vars_num <- names(customers)[sapply(customers, class) == "Date"]
    paste0(sapply(
      vars_num,
      FUN = function(i)
        paste(i, "=", input[[i]])
    ),
    collapse = '\n')
  })
}
shinyApp(ui = ui, server = server)
}
}
