% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinputs.R
\name{autoinputs}
\alias{autoinputs}
\title{autoinputs}
\usage{
autoinputs(.df, .dec_places = 0)
}
\arguments{
\item{.df}{Dataframe with N variables (date, factor or numeric types are supported)}

\item{.dec_places}{Decimal places for the numerical sliders}
}
\value{
Input buttons for all variables.
}
\description{
Automatically genrates inputs for all variables in a dataframe.
}
\details{
Based on the columns on a dataframe, this function generates all the inputs.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
load('data/customers.rda')
ui <- fluidPage(fluidRow(
  column(3,
         h3('Inputs based on df'),
         autoinputs(.df=customers)
  ),
  column(3,
         h3('Outputs based on inputs'),
         verbatimTextOutput(outputId = 'values')
  )
))
server <- function(input, output) {
  output$values <- reactive({
    paste0(sapply(
      names(customers),
      FUN = function(i) paste(i, "=", input[[i]])),
      collapse = '\n')
  })
}
shinyApp(ui = ui, server = server)
}
}
