% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autooutput_df.R
\name{autooutput_df}
\alias{autooutput_df}
\title{Title autooutput_df}
\usage{
autooutput_df(.df, .inputs, .dates_as_str = FALSE)
}
\arguments{
\item{.df}{The original dataframe}

\item{.inputs}{Input from the ui}

\item{.dates_as_str}{dates as character}
}
\value{
A dataframe based on the input values generated with 
        shinymatic::autoinputs()
}
\description{
Based on the values of the inputs selected on the UI, 
             a reactive dataframe is generated
}
\details{
This df could be used as the input parameters for a machine learning
         model, where another output is the predicted value based on the 
         dataframe.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
ui <- fluidPage(
  column(3,  h3('Inputs based on df'),
         autoinputs(.df=customers)),
  column(6, h3('Outputs based on inputs'),
         tableOutput(outputId = 'data_test'))
)
server <- function(input, output) {
  data_preds <- reactive({
    autooutput_df(.df=customers, .inputs=input)
  })
  output$data_test <- renderTable ({
    data_preds()
  })
}

shinyApp(ui = ui, server = server)
}
}
