% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\docType{data}
\name{customers}
\alias{customers}
\title{Customers: a random dataframe with columns of different types}
\format{
A data frame with 10 rows and 9 variables:
\describe{
  \item{birthdate}{Birth date of the customer}
  \item{country}{Country of the customer}
  \item{category}{Category of the customer}
  \item{amount_tot}{Total amount spent by customer}
  \item{amount_avg}{Average amount spent by customer}
  \item{first_date}{First day of customer purchase}
  \item{last_date}{Last day of customer purchase}
}
}
\source{
Random data generated
}
\usage{
customers
}
\description{
The dataframe was generated with customers_data.R 
             (shinymatic::customers_data)
}
\keyword{datasets}
