% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers_data.R
\name{customers_data}
\alias{customers_data}
\title{customers_data}
\usage{
customers_data(.N_customers)
}
\arguments{
\item{.N_customers}{Number of random observations to generate}
}
\value{
A dataframe with .N_customers observations and 7 variables
}
\description{
Generate N samples of random customer data
}
\examples{
customers_data(.N_customers=1000)
}
