% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_categorical.R
\name{autoinput_categorical}
\alias{autoinput_categorical}
\title{autoinput_categorical}
\usage{
autoinput_categorical(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be factor)}
}
\value{
Select input for categorical variables
}
\description{
Automatically genrates select inputs for all factor columns in a dataframe.
}
\details{
Based on the factor columns on a dataframe, this function generates all the select inputs. 
         The possible values correspond to the levels in each factor column.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
load('data/customers.rda')
ui <- fluidPage(
  h3('UI (inputs)'),
  autoinput_categorical(.df = customers),
  h3('Server (output)'),
  verbatimTextOutput(outputId = 'values')
)
server <- function(input, output) {
  output$values <- reactive({
    vars_cat <- names(customers)[sapply(customers, is.factor)]
    paste0(sapply(
      vars_cat,
      FUN = function(i)
        paste(i, "=", input[[i]])
    ),
    collapse = '\n')
  })
}
shinyApp(ui = ui, server = server)
}
}
