% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_categorical.R
\name{autoinput_categorical}
\alias{autoinput_categorical}
\title{autoinput_categorical}
\usage{
autoinput_categorical(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be factor)}
}
\value{
Select input for categorical variables
}
\description{
Automatically genrates select inputs for all factor columns in a dataframe.
}
\details{
Based on the factor columns on a dataframe, this function generates all the select inputs. 
         The possible values correspond to the levels in each factor column.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
df <- data.frame(
  var_num_1 = c(1, 2, 3),
  var_num_2 = c(6, 5, 4),
  var_cat_1 = factor(c('a', 'b', 'c'))
)
ui <- shiny::fluidPage(
  h3('UI (inputs)'),
  autoinput_categorical(.df = df),
  h3('Server (output)'),
  verbatimTextOutput(outputId = 'values')
)
server <- function(input, output) {
  output$values <- reactive({
    vars_cat <- names(df)[sapply(df, is.factor)]
    paste0(sapply(
      vars_cat,
      FUN = function(i)
        paste(i, "=", input[[i]])
    ),
    collapse = '\n')
  })
}
shiny::shinyApp(ui = ui, server = server)
}
}
