% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_date.R
\name{autoinput_date}
\alias{autoinput_date}
\title{autoinput_date}
\usage{
autoinput_date(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be date)}
}
\value{
dateInput() for all the date variables.
}
\description{
Automatically genrates date inputs for all date variables in a dataframe.
}
\details{
Based on the date columns on a dataframe, this function generates all the date inputs. 
         The min/max value on the dataframe is the min/max value on the date input.
}
\examples{
## Only run this example in interactive R sessions
#' if (interactive()) {
library(shiny)
library(shinymatic)
df <- data.frame(
  var_num_1 = c(1, 2, 3),
  var_num_2 = c(6, 5, 4),
  var_cat_1 = c('a', 'b', 'c'),
  var_date_1 = as.Date(c('2022-01-01','2022-04-02','2022-05-10')),
  var_date_2 = as.Date(c('2021-12-01','2022-03-02','2022-06-20'))
)
ui <- shiny::fluidPage(
  h3('UI (inputs)'),
  autoinput_date(.df = df),
  h3('Server (output)'),
  verbatimTextOutput(outputId = 'values')
)
server <- function(input, output) {
  output$values <- reactive({
    vars_num <- names(df)[sapply(df, class) == "Date"]
    paste0(sapply(
      vars_num,
      FUN = function(i)
        paste(i, "=", input[[i]])
    ),
    collapse = '\n')
  })
}
shiny::shinyApp(ui = ui, server = server)
}
}
