% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_numerical.R
\name{autoinput_numerical}
\alias{autoinput_numerical}
\title{autoinput_numerical}
\usage{
autoinput_numerical(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be numeric)}
}
\value{
SliderInputs for all the numeric variables.
}
\description{
Automatically genrates slider inputs for all numeric variables in a dataframe.
}
\details{
Based on the numeric columns on a dataframe, this function generates all the slider inputs. 
         The min/max value on the dataframe is the min/max value on the slider.
}
\examples{
library(shiny)
library(shinydfinputs)
df <- data.frame(
  var_num_1 = c(1,2,3),
  var_num_2 = c(6,5,4),
  var_cat_1 = c('a','b','c'))

ui <- shiny::fluidPage(
  autoinput_numerical(.df=df),
  textOutput(outputId = 'values')
)

server <- function(input, output) {
  output$values <- reactive({
    vars_num <- names(df[,unlist(lapply(df, is.numeric))])
    paste0(sapply(vars_num, 
                  FUN=function(i) paste(i,"=", input[[i]])), 
           collapse = ', ')
  })
}

shiny::shinyApp(ui = ui, server = server)
}
