% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinput_numerical.R
\name{autoinput_numerical}
\alias{autoinput_numerical}
\title{autoinput_numerical}
\usage{
autoinput_numerical(.df, .dec_places = 0)
}
\arguments{
\item{.df}{Dataframe with N variables (one or more should be numeric)}
}
\value{
SliderInputs for all the numeric variables.
}
\description{
Automatically genrates slider inputs for all numeric variables in a dataframe.
}
\details{
Based on the numeric columns on a dataframe, this function generates all the slider inputs. 
         The min/max value on the dataframe is the min/max value on the slider.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
library(shiny)
library(shinymatic)
load('data/customers.rda')
ui <- fluidPage(
  h3('UI (inputs)'),
  autoinput_numerical(.df = customers),
  h3('Server (output)'),
  verbatimTextOutput(outputId = 'values')
)
server <- function(input, output) {
  output$values <- reactive({
    vars_num <- names(customers)[sapply(customers, is.numeric)]
    paste0(sapply(
      vars_num,
      FUN = function(i)
        paste(i, "=", input[[i]])
    ),
    collapse = '\n')
  })
}
shinyApp(ui = ui, server = server)
}
}
