% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoinputs.R
\name{autoinputs}
\alias{autoinputs}
\title{autoinputs}
\usage{
autoinputs(.df)
}
\arguments{
\item{.df}{Dataframe with N variables (date, factor or numeric types are supported)}
}
\value{
dateInput() for all the date variables.
}
\description{
Automatically genrates inputs for all variables in a dataframe.
}
\details{
Based on the columns on a dataframe, this function generates all the inputs.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
load('data/df.rda')

ui <- shiny::fluidPage(fluidRow(
  column(3,
         h3('Inputs based on df'),
         autoinputs(.df=df)
  ),
  column(3,
         h3('Outputs based on inputs'),
         verbatimTextOutput(outputId = 'values')
  )
))

server <- function(input, output) {
  output$values <- reactive({
    paste0(sapply(
      names(df),
      FUN = function(i) paste(i, "=", input[[i]])),
      collapse = '\n')
  })
}

shiny::shinyApp(ui = ui, server = server)
}
}
